Added support for variable-size colormap (up to 8192 colors) and "log log" smoothing (see http://linas.org/art-gallery/escape/smooth.html).
--klm

+++ plug-ins/FractalExplorer/Dialogs.h
@@ -11,7 +11,7 @@ void                dialog_create_value(char *title, GtkTable * table,
 					scaledata * scalevalues);
 
 void                dialog_create_int_value(char *title, GtkTable * table, 
-					    int row, gdouble * value,
+					    int row, int * value,
 					    int left, int right, 
 					    const char *desc, 
 					    scaledata * scalevalues);
